kali-agent/requirements.txt
aiogram>=3.4
openai>=1.30
aiosqlite
pyyaml
pydantic
python-dotenv
